Use float16 codec to fit HeartMuLa 3B on 24GB GPUs
float32 codec used ~5GB VRAM, leaving only 679MB free on a 24GB
RTX 4090 — not enough for generation. float16 codec halves that
with imperceptible quality difference (codec is just the audio
encoder/decoder, not the quality-critical LLM backbone).

Frees ~11GB VRAM: 679MB free → 12.7GB free.

+++ tuneforge/generation/heartmula_backend.py
@@ -77,7 +77,7 @@ class HeartMuLaBackend:
                 },
                 dtype={
                     "mula": torch.bfloat16,
-                    "codec": torch.float32,
+                    "codec": torch.float16,
                 },
                 version=self._version,
                 lazy_load=False,
